Update resume.md
Made a handful of changes here... take what's useful/leave what isn't.  Also happy to discuss next week.

The two "big" changes I made were to replace the "Professional Summary" header with a summary of your role (too valuable a space to waste), and removed the "Leadership Skills" section, which I don't typically see as its own section (should be captured in the initial paragraph or reflected in work experience).

Smaller changes: adjusted strengths & expertise so each column is a "theme" (team leadership, technical skills, product/project skills); re-ordered and re-worded some of your bullets; broke out a "Languages & Frameworks" section for each role (not committed to this, but it was a way to capture the tools you worked with in a consistent way across roles).

Thanks,
Aaron